Default sleep timer consistent as a shortcut
Prior to this patch, changing the default sleep timer setting when used
as a shortcut did not update the underlying sleep timer value until the
configuration file was re-read.

The only downside is removing the sleeptimer_duration_cb function means
that the timer is restarted even if you do not change the default
duration.

Change-Id: I178f50cf8658dc64ecb09d6b30223aee0c6d7ad1