Use crc32 of filename to resume tracks
As well as using an index, which breaks when a file is added or
removed, use the crc32 of the filename. When the crc32 check passes the
index is used directly. When it fails, the slow path is taken checking
each file name in the playlist until the right crc is found.

See http://www.rockbox.org/tracker/6411

Change-Id: I156f61a9f1ac428b4a682bc680379cb6b60b1b10